requirements.txt
kaggle==1.5.12 
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-file-share==12.10.1
logging==0.4.9.6
json==1.6.3